backend

+++ ms-chat/src/main/java/com/example/ms_chat/repository/ChatMessageRepository.java
@@ -6,7 +6,7 @@ import com.example.ms_chat.entity.ChatMessage;
 
 import java.util.List;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Integer> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
     // Métodos para obtener el historial de chat entre dos usuarios
     List<ChatMessage> findBySenderIdAndRecipientIdOrderByTimestampAsc(String senderId, String recipientId);
     List<ChatMessage> findBySenderIdInAndRecipientIdInOrderByTimestampAsc(List<String> ids1, List<String> ids2);
